Kc1

+++ public/js/appointments.js
@@ -1,15 +1,17 @@
 const deleteAppt = async (event) => {
-  event.preventDefault();
-  const response = await fetch(`/api/appointments/remove/${$(event.target).attr("data-id")}`, {
-  method: 'DELETE'
-});
+  if(event.target.hasAttribute('data-id')){
+    const apptId = event.target.getAttribute('data-id');
+    const response = await fetch(`/api/projects/${apptId}`, {
+      method: 'DELETE',
+    });
     if (response.ok) {
-        document.location.reload();
-      } else {
-        console.log(response.statusText)
-        alert('Could not delete appointment.');
-      }
+      document.location.replace('/appointments');
+    } else {
+      console.log(response.statusText)
+      alert('Could not delete appointment.');
+    }
   }
+};
 
   $('#confirm-btn').click(async (event) => {
     event.preventDefault();
